+++ css/style.css
@@ -1014,6 +1014,11 @@ fieldset {
   opacity: 0.6;
 }
 
+.ad-form__error-text {
+  font-size: 0.75em;
+  color: #ff5635;
+}
+
 .visually-hidden {
   position: absolute;
